Updated router controller
